requirements.txt
streamlit~=1.48.1
pandas~=2.3.1
numpy~=2.3.2
mysql-connector-python~=9.3.0
PyYAML~=6.0.2
plotly~=6.3.0
matplotlib~=3.10.5
streamlit-aggrid~=1.1.8.post1
langgraph~=0.5.4
langchain-core~=0.3.70
langchain-openai~=0.3.27
langchain~=0.3.26
dotenv~=0.9.9
python-dotenv~=1.1.1
protobuf~=6.32.1
langchain-community~=0.3.27
tiktoken~=0.9.0
pinecone~=7.3.0
openai~=1.95.1
tldextract~=5.3.0
requests~=2.32.5
PyPDF2~=3.0.1
beautifulsoup4~=4.13.4
tenacity~=9.1.2
scikit-learn~=1.7.1
tabulate~=0.9.0
SQLAlchemy~=2.0.41